+++ userprofile/views.py
@@ -66,36 +66,8 @@ class UserDetailView(LoginRequiredMixin, LoggedInPermissionMixin, TemplateView):
     template_name = 'userprofile/detail.html'
     permission_required = 'userprofile.view_detail'
 
-    def _get_user_settings(self):
-        """Return user setting values"""
-        plugins = plugin_api.get_active_plugins(
-            plugin_type='project_app'
-        ) + plugin_api.get_active_plugins(plugin_type='site_app')
-        for plugin in plugins + [None]:
-            if plugin:
-                name = plugin.name
-                s_defs = app_settings.get_definitions(
-                    APP_SETTING_SCOPE_USER, plugin=plugin, user_modifiable=True
-                )
-            else:
-                name = 'projectroles'
-                s_defs = app_settings.get_definitions(
-                    APP_SETTING_SCOPE_USER,
-                    plugin_name=name,
-                    user_modifiable=True,
-                )
-            for s_def in s_defs.values():
-                yield {
-                    'label': s_def.label or f'{name}.{s_def.name}',
-                    'value': app_settings.get(
-                        name, s_def.name, user=self.request.user
-                    ),
-                    'description': s_def.description,
-                }
-
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context['user_settings'] = list(self._get_user_settings())
         context['add_emails'] = SODARUserAdditionalEmail.objects.filter(
             user=self.request.user
         ).order_by('email')